+++ ReleaseNotes.md
@@ -32,6 +32,8 @@
 * TGR-1842: Workflow Ui: fixed an issue where clicking a tree element in the RbelPath inspect dialog would prefill the
   dialog with an incorrect expression.
 * TGR-1816: Tiger Proxy: it is now possible to use a direct binary proxy for pop3 and smtp without TLS.
+* TGR-1864: Fixed a mesh setup racing condition between a message being removed from history of the downstream proxy
+  before the following message arrives that waits for the previous message to have be processed.
 
 # Release 3.7.9